Replace future with six for python2/3 compatibility.
future was an undeclared (on setup) dependency.  It was not a core
python package (different from __future__).  We replace it with six
because it does the same thing and we are already dependent on it from
serpent (which is a Pyro4 dependency).  It also seems to implement a
slightly more efficient version of iteritems by selecting the
implementation at import time.