Standardise method to load shared libraries
Always use the CDLL and its subclasses for the sake of "one-way".
Means less work when monkey-patching ctypes for testing and building
of documentation.

Use os.name instead of platform.architecture to decide on use of
WinDLL, which is the same method used in CPython.

Remove duplicated code in SDK3.py (patch forwarded to PyME too).