python 3 support for lasers by ensure serial commands are bytes
python 3 text strings are not bytes and need to be converted properly.
pyserial is also expecting bytes always so handle all communication
as bytes.  Decode as utf-8 only when returning back to the user or
sending to the logger with a format string.

The logger itself does not care if it's bytes or str, but if it's
a format string and one of the arguments is bytes, then that gets
formatted with b'...' on the final message.

Also pass format string and args to the logger methods instead of the
formatted string do that the formatting only happens if that log level
is enabled.