SerialDeviceMixIn: initial work on a Mixin for devices controlled over serial
There's some code duplication between the laser devices which all have
a serial connection until now.  The main example is the lock_comms
decorator in the sapphire and cobolt module which don't actually flush
input like they say they do in the docs, versus _flush_buffer in the
deepstar module which does flush.

There's ground to do more refactoring in the future, but for now, we
just ensure that the comm locking, and passthrough to the serial
connection are in this MixIn.

Also remove the unused private method _read().