Remove Pyro4.expose everything by disabling REQUIRE_EXPOSE (issue #49)
Pyro4 requires to mark methods and classes that should be exposed so
shit Pyro4.expose all over the code.  And sometimes we forget to do it
in a few places.  Since we effectively manually expose everything, we
might as well disable that requirement and automatically expose
everything.

This is a first step on decoupling Pyro4 from the devices interfaces.