microscope.testsuite.libs: remove module and adjust setup.py for it
The microscope.testsuite.libs module provided a mock of all C
libraries and functions required to import all microscope modules.
These are required by setup.py when building the documentation and
would be needed for testing devices one day.  However, these two needs
are quite different.

For building documentation, we only need to be able to load libraries
and access the functions.  A standard MagickMock is enough.  So move
this into setup.py.  This also solves the previous problem that a
microscope module (microscope.testsuite.libs) was needed to run the
microscope setup.py script.

For running the tests, and we don't yet have tests for devices that
need this libraries, we will need a functional mock of the libraries.
When it does become needed, we will need a better mock of C functions
type that handle the type checking and conversion.  We will probably
want to keep each of them on its own file together with its tests, and
not all in one file.  This solution will not be suitable anyway.