test_client: annotate test pyro objects with @expose
The tests in test_client were failing if ran before importing
microscope.deviceserver since 708c7b50511a because the objects
members were not being exposed.  Only on deviceserver were we
changing the global config.  This change exposes the members
required for the test so that it passes independently of our own
Pyro4 config changes.