Fixed Thorlabs filter wheel and updated interface.
The wheel position and filters were exposed as settings, but this
complicates things needlessly: clients should just call get_position,
set_position and get_filters, directly. Nothing used the 'filters'
setting, so I've removed it now. I've marked the 'position' setting as
deprecated, as it is currently used by cockpit. Eventually, it should be
removed.

The 'filters' argument is now an optional keyword for all filterwheels. It
can be a list, or a dict mapping possition to filter, where each entry is
either a description string, or a tuple of description and some value
(e.g. wavelength, neutral density, ...).
The FilterWheelBase now offers another keyword, 'positions', that may be
used to set the maximum number of positions on a wheel instead of
providing filter definitions. Where the number of positions is fixed by
the filter wheel model (often the case), then the maximum number of
positions can be set by assinging _positions in the class definition, as
has been done here for the Thorlabs FW102C and FW212C wheels.

Some filter wheels have manual controls. I considered having a polling
thread to regularly query the wheel position and thereby catch manual
position changes. However, filterwheels currently have no mechanism to
notify the client of a change, so we leave it to the client to ensure they
query the wheel position as needed, either before actions where the
wheel's position is critical, or regularly, as required.