Register LinkamBase as a FloatingDeviceMixin instead of LinkamCMS
FloatingDeviceMixin is an ABC that requires get_id.  This is
implemented in LinkamBase and not on LinkamCMS.

Also, mixins need to appear before base classes to have the MRO
done in the right order.  LinkamCMS had the FloatingDeviceMixin
last on the MRO to work around this (list FloatingDeviceMixin
before LinkamBase would have get_id call the empty implementation
on the mixin only).  This change also gets a correct MRO.