Device: raise an exception if there's unexpected arguments (issue #84)
Do not accept *args and **kwargs on the base Device class.  These were
unused arguments anyway.  This will cause an exception to be raise if
a device is constructed with arguments that are not used by any of its
parent classes.  This prevents the user from accidentally constructing
a device with incorrect arguments because of a typo.