Device: replace Device._logger with module loggers (issue #110)
Each device instance has its own logger.  This was meant to help in
filter the log records by device.  This was achieved by having
DeviceServer get the logger with that device name and have records
from it written to a properly named file.  However, this meant that
log record from other loggers were discarded and has since been
changed.  This means that there's no longer need for each Device to
have its own instance and can instead used module loggers as is
recommended in the documentation and is standard in other packages.

This commits removes the Device._logger attribute and replaces all its
use with loggers from the individual modules.