Remove @Pyro4.behavior(instance_mode='single') (issue #49)
Setting Pyro's instance mode behaviour to 'single' means that a single
instance of the decorated class is created and used for all methods
calls.  This is the same as creating and registering an instance with
the Pyro daemon which is what deviceserver already does.

This commit removes those decorators which are doing nothing.  This
makes the concrete classes independent of Pyro.