deviceserver: force multiprocessing spawn start method in MacOSX (issue #106)
MacOSX High Sierra changed behaviour of fork in Objective-C code.
+initialize methods are not supposed to be called between fork() and
exec().  However, Python's multiprocessing fork start method uses
fork() without exec().  Python has fixed this in version 3.8 by
changing the default start method in MacOSX to spawn.  This commit
does the same.  The workaround should be removed once we are
dependent in Python 3.8 or later so that a user can easily change
it to something else.