microscope.controllers.prior: add support for Prior ProScanIII controller.
ProScanIII is the concrete class implementing the controller interface
for the Prior ProScanII device.  For now it can only control filter
wheels.  There is work in progress to add support for stages but it is
dependent on defining a stage interface (issue #99).  We don't add
support for light sources because we don't have access to them.  And
don't add support for shutters because there's not even work in
progress to add an ABC for them.

_ProScanIIIFilterWheel is the concrete class implementing the filter
wheel interface.  It was tested with a HF108A and a HF110A filter
wheel.

_ProScanIIIConnection wraps the serial connection and its commands so
they can be shared between future classes that add support for other
device types.  Also handles synchronization of communication from
different threads.