Add support for Lumencor Spectra III Light Engine (issue #109)
The Lumencor Spectra III light engine is a device that controls
multiple light sources, not necessarily all lasers.  This is still
modelled as a ControllerDevice controlling LaserDevices.

This was implemented for a Spectra III 8-LCR-SA model but should work
with any Spectra III model.  From the documentation, it should be
pretty easy to make it work for any lumencor light engine.