Make the Setting class private to enable refactoring of Settings (issue #125)
Managing different types of Settings may be achieved more flexibly
and simply through the use of subclasses of Setting and the use of a
factory pattern or helper method for their creation.  So make the
`Setting` class and the `Device.settings` attribute private so that
the `Device` methods become the only sactioned method to get to them.

* microscope/cameras/atmcd.py: use the `*_setting` methods instead of
  manipulating `self.settings` directly.
* microscope/devices.py: mark `Setting` class and `Device.settings`
  attribute as private.
* microscope/testsuite/test_setting.py: Setting class is now _Setting.
* NEWS: make note of backwards incompatible change.