win32: explain what's going on when importing the win32* modules
Mick reports that importing the win32* modules may fail because they
import win32api which is a pyd file.  We catched this by trying to
import win32api even though we were not actually making use of it but
then code analysers complain about unusued import.

This commit makes it so that we import what we actually use and
documents why we are manipulating PATH (which was previously not clear
anyway).