Pass the % format args to loggers instead of doing formatting ourselves.
We call Logger.debug (or any other logging level) with a message but
we don't know if that message will be emitted.  If it never gets
emitted, it's a waste to format the message.

So instead of doing the % formatting format ourselves with %, this
commit passes the unformatted message and the % args to the Logger
which delays the formatting until it knows the message will actually
be emitted.