Added error tolerance to get_all_settings.
get_all_settings reports the values of problem settings as None.
Previously, we considered adding a new entry, __errors__, to the returned dict, but this is adding an item that is not a setting to a dict where everything else is a setting, and would require clients to handle this key as a special case.