Device.get_all_settings: return None for settings where get() fails
Use None as value to report those settings that failed to get() its
value.  Previously, we considered adding a new entry, __errors__, to
the returned dict, but this is adding an item that is not a setting to
a dict where everything else is a setting, and would require clients
to handle this key as a special case.