AlpaoDeformableMirror._err_msg: remove internal attribute.
There was a time when AlpaoDeformableMirror was checking the error
stack for a message.  To avoid constantly recreating the message
buffer, there was an instance attribute with a C string buffer that
was constantly emptied and reused.

The error stack is now only checked if a function does not succeed so
this extra work is pointless.