XimeaCamera: connect by serial number instead of device ID.
In the case of a multi camera setup, the device ID for Ximea cameras
is defined after the first xiAPI call.  This means that we can't use
the device ID in the XimeaCamera initialiser to set the camera that
should be used.

This replaces the dev_id argument with serial_number which should be
unique and known by the user ahead of time.