XimeaCamera: replace `_triggered` attribute with Camera.get_image timeout.
If camera is on software trigger mode, sending triggers would set the
`_triggered` attribute to `True`.  The fetch loop would notice that it
was `True`, get one image, and then set it to `False`.  If more than
one software trigger were sent before the image was read, only the
first image would be read.  A later trigger would then read the second
image on the queue buffer instead of the new trigger.

Because currently XimeaCamera has no real way to change trigger mode,
it was set to hardware trigger mode and relying on the fact that the
camera would still answer to software triggers in that mode.  This
meant that the `_triggered` code path was actually never used and so
probably no one has been affected by this bug.

This commit completely removes the `_triggered` attribute with always
relying on `Camera.get_image` to timeout if there is no image
available.