XimeaCamera: update interface to make use of Binning class (issue #128)
The CameraDevice.get_binning, and CameraDevice.set_binning methods
used to take/return two arguments.  They have been changed to use
Binning instances with commit e70555f39dc4 but the Ximea interface has
never been updated.

This commits does that and prepares the code to actually add support
for binning.  This seems straightforward with xiapi but we don't have
a camera that actually supports binning to test it and we are afraid
to distribute something that does not work.