_Setting.describe: remove unnecessary conversion of dtype to str.
Before b9e5c1c91053, the _Settings.dtype value ancestor (this predates
the Setting class) could be a string or an actual type.  The type was
then converted to str because "can't serialize types".

Since b9e5c1c91053, _Settings.dtype is always the string with the
type name.  The conversion to str is no longer required.  This
commit removes the conversion.