Device.add_setting: remove support to specify dtype by type.
When adding a new setting it is possible to specify dtype with the
type instead of the type name, e.g., int instead of 'int'.  However,
this is undocumented, does not work with enum types and may not work
with future dtypes, seems to add nothing other than another way to do
the same thing.

The first value of DTYPES values was the type description which was
added in b9e5c1c91053 to support Python 2 and 3 (issue #73), also only
required when Setting.dtype was a type.

This commit 1) removes support to specify dtype via type; 2) replaces
its few usage with type names; and 3) removes type description from
DTYPES values() first element.