CoolLED: check if channels are "selected" and warn if not (issue #138)
CoolLED channels have a selected/unselected state and an on/off state.
If the channel is unselected then it can only off and turning it on
will not work.  We could programatically select it as well but maybe
we shouldn't and we're not sure what will happen in a pE-4000.

This commits will emit a warning about the state and suggest that the
users enables it on control pod.