AndorAtmcd: fix BaselineClamp, EMAdvanced, and HighCapacity (fixup 9b58109f)
It used to be possible to add settings by creating a Setting instance
and directly inserting it into the `Device._settings` dict.  When
using the Setting constructor, the `values` (5th argument) was
optional but when using `Device.add_setting` it is not optional.

Commit 9b58109fcbf1 replaced all inserts to `Device._settings` with
`Device.add_setting`.  Settings of dtype bool don't seem to make use
of the `values` argument and were relying on the default of `None`.

This commit specifically sets `values` as `None`.  An alternative fix
could be to make `values` optional in `Device.add_setting`.