XimeaCamera: implement get_trigger_type method (cockpit#561)
CameraDevice.get_trigger_type is not an abstract method so XimeaCamera
does not implement it, instead users are expect to use the
TriggerTargetMixIn methods.  However, cockpit can't use that.

This commit implements get_trigger_type method so that Ximea cameras
can be used with hardware triggers (and therefore experiments) in
Cockpit.