deviceserver: check FloatingDevice UID even when only one is expected (#153)
FloatingDevices can only be identified after having been constructed
(the constructor will get return any of the devices).  DeviceServer
only checked the UID if there was more than one device listed on the
config.  This meant that if the user had only one device, it didn't
had to put the device UID correctly.  However, it also meant that if
the user had two devices connected but only one on the config file it
could get the other one without any error because the UID was never
checked.

This commit ensures that UID of FloatingDevices are always checked and
adds a test case for the issue.