deviceserver: support function instead of type in device definition (#154)
A device definition includes the type of device to be created and the
arguments to construct it.  This is aimed at serving only one device
per Pyro server and process and prevents the serving of multiple
devices on the same Pyro server (which is desirable when there's data
transfer between devices).  Even with only one device, the separation
of kwargs and class makes configuration is a bit too verbose.

This commit adds the ability to pass a function that returns multiple
devices in the device function that constructs device definition, and
the required logic on DeviceServer.  As a side effect, this feature
also adds support to specify the Pyro4 URI.