microscope.gui: rewrite for qt and new microscope-gui program.
This commit rewrites the whole microscope.gui module to use qt instead
of tkinter.  While originally this module was developed with tkinter
because it wouldn't require extra dependencies, tkinter is just too
limiting and undocumented and there are now multiple qt interfaces in
python that are easy to install.

This commit also adds a new setuptools entry point, microscope-gui,
which will display the widget given a Pyro URI.