LaserDevice: replace power methods to use fractional values instead of mw (#155)
This commit removes the methods that handle power in LaserDevice in
physical units (mW) and replaces them with methods to set relative
laser power (in the 0-1 range).  In the cases where the device does
not have commands to set relative laser power and the max laser power
in mW is required internally, the get_max_power_mw method was replaced
with an internal attribute set during construction.