FilterWheelBase: remove the whole filters stuff from class.
This commit removes the filters argument from the class constructor
and the `get_filters` method.  This also removes the implicit ability
of guessing number of positions from the number of filters but also
makes it clear what happens when both are specified.

The reason for this change is that which filter is at each position is
not a low level operation suitable for this.  Instead, it should be
implemented by the end user interface of the microscope, e.g., cockpit
which already does it.  As far we know, no one is using this ability
anyway.