FilterWheelBase: remove default of zero positions.
Having a default value for number positions does not make sense
because there's no reason that a specific value is more likely.
Having a default value of zero more so, since it's not even a valid
number of positions.

This commit removes the default number of positions in FilterWheelBase
and requires subclasses to pass it as argument instead of assuming a
class attribute which causes issues for static code analysis.