DeformableMirror: new _do_apply_pattern method for device specific code.
All implementations of `apply_pattern` validate the pattern before applying it.
We also plan on adding a check for trigger type (see issue #61).  So we should
have this method split in two, one for common code, and other for device
specific code.

This commit makes `DeformableMirroe.apply_pattern` no longer an abstract method
with the common code.  It then calls the new abstract method `_do_apply_pattern`
for the device specific code.