Move and rename multiple modules, classes, and functions.
This commit standardises the ABC class names.  We had things like CameraDevice,
DeformableMirror, and FilterWheelBase so the Device and Base suffixes were
removed for non-mixin classes, i.e., CameraDevice is renamed to Camera,
ControllerDevice is renamed to Controller, FilterWheelBase is renamed to
FilterWheel, LaserDevice is renamed to Laser, and StageDevice is renamed to
Stage.

In the mixin classes, some had the suffix MixIn while other had Mixin.  This
commit sets on the Mixin capitalisation, i.e., SerialDeviceMixIn is renamed to
SerialDeviceMixin, and TriggerTargetMixIn is renamed to TriggerTargetMixin.

This commit also moves all the abstract base classes and functions meant to be
used by them (effectively, most of the microscope.devices module) to a new
module named microscope.abc.  This should not affect any user since users don't
interact with these classes and functions directly.

This commit moves the AxisLimits, Binning, ROI, TriggerMode, and TriggerType
classes, which the user will interact directly, to the root of the package, the
microscope module.  The TRIGGER_* enum is kept in the microscope.abc module
since the plan is to replace them with TriggerMode and TriggerType.

This commit renames deviceserver module to device_server.

This commit moves the device function from the microscope.devices module to the
microscope.device_server module.

This commit adds microscope.devices and microscope.device_server modules that
import the previous attributes from the new locations with the old names for
backwards compatibility.