Update copyright notice and standardise on their style.
While all files are under GPLv3+, different files set it with slightly different
styles which make it more difficult to automate changes to them.

This commit standardises on a style, starts on line three, after the shebang and
an empty line, uses two #, includes name of the project in the copyright notice
as per FSF recommendation, uses <> for the email address, and sorts copyright
holders.

This commit also updates email addresses, names of copyright holders (in some
were David Pinto and others David Miguel Susano Pinto), and updates the
copyright notice to year 2020 (with the exception of SDK3 and SDK3Cam original
copyright which are kept with the year when we forked it).