Raise TypeError and ValueError built-in exceptions instead of Exception (#134)
We mostly throw the generic `Exception` but plan on having microscope specific
exceptions.  But first, there are cases where we can use specific built-in
specific exceptions.

This commit replaces raising of Exception with TypeError for cases where the
function is called incorrect, be it wrong type, wrong number, or wrong
combination.  It replaces raising Exception with ValueError when an argument has
the right type but an invalid or unsupported value (but not where the problem is
a specific device does not support some feature).