Camera: declare Camera ABC as subclass of TriggerTargetMixin (#159)
All concrete implementations of the Camera ABC already implement the
TriggerTargetMixin interface.  This commit remove it from the
individual concrete classes and declare it on the Camera ABC.

This commit also moves up the declaration of the TriggerTargetMixin in
the microscope.abc module so that it can be used by the Camera ABC.

This commit removes the soft_trigger method from the Camera ABC since
the TriggerTargetMixin already declares a trigger method.  The current
concrete implementations of Camera keep the soft_trigger method for
now for backwards compatibility.