DeviceServer: do not initialize() if getting devices from function (#160)
This commits changes DeviceServer so that it no longer calls initialize() on the
devices it they were constructed from the passed function.

A DeviceServer can be using a class to construct the device or a function that
will construct any number of devices.  If using a class, then DeviceServer calls
initialize automatically, the device will be useless otherwise.  However, using
a function provides the user the ability to change default settings of the
served device.  This requires the device to be initialised.  Since there's no
methods to check if a device has been initialised (not sure if it should), and
since most calls to initialize() will fail if the device already has been
initialised, we change to now require that a function returns devices that
already are initialised.  Since the ability pass a function is new, this is not
backwards incompatible.