Camera: remove "binning" setting from Camera ABC (#161)
We have set_binning method, the "binning" setting on the Camera ABC, and each
concrete camera class has its own binning related method.  The first two are
redundant.  Since we want settings to be direct and specific to each device, and
since the ABC requires a set_binning method, we remove the "binning" setting.