microscope.abc.Laser: renamed to LightSource.
The Laser ABC is actually being used for other light sources, such as
CoolLED, Zaber LEDs, and Lumencor controllers.  So we rename this class
to make it clear.  Because this only changes the name of the ABC, there
is no change on the actual devices behaviour, so no user changes.