Camera.get_sensor_temperature: remove method from ABC.
The `Camera.get_sensor_temperature` is not an abstract method and
there's nothing at the moment that uses it (not even in cockpit).  The
ximea camera also has more than one temperature sensor.

We do want to have a way to get temperature from the camera but this
method does not look like the way to go.  Maybe in the future have a
get_sensors method (for multiple sensors), but for now they should use
the settings dict.