Replace _on_shutdown, _on_disable, _on_enable with _do_* equivalents.
On some cases we have _do_x as the abstract method that actually does
the device specific code.  In other cases we have _on_x.  This commit
standardises on _do_x.  Either work well, but _on_x would suit better
to the method that is called in response, i.e., after, an event and
not the method that is called to do that actual thing.