Device.make_safe: remove method which is effectively the same as disable.
This method is not an abstract method and when it was implemented it
was effectively the same as disable or abort.  It was only implemented
on AndorSDK3, PVCamera, and XimeaCamera and in those cases it simply
called `abort`, the same that `disable` does.  No other class
implemented make_safe, not even the AndorAtmcd camera class.

This commit removes the Device.make_safe method, which can be replaced
with calling disable.