LightSource: implement TriggerTargetMixin for all light sources (#159)
We need a way to configure how the light source emits light.  At the
moment, after being enabled, some light sources (Deepstar and Obis)
will wait for the TTL signal to start emit light, while the others
start emitting light as soon as enabled is called.  We want to support
hardware trigger for most of those.

This commit makes LightSource ABC subclass from the TriggerTargetMixin
and adjusts the concrete implementations with their current behaviour.
On future commits, more trigger types should be added to those devices
that support it.