Device: don't have initialize as an abstract method.
The base device class has an abstract method initialize which is
supposed to be for initialise connection with the device.  In practice
having a separate initialize method is only to used to attempt to get
the device back into working order after it has crashed and rebooted
(see issue #129).

Having a separate initialize method that is required means a separate
method call that always needs to be called to use the device.  This
change documents that __init__ of concrete implementations must call
initialize so that the device is wokring order after having been
constructed.  Some concrete implementations may keep a separate
initialize method but it's not required anymore, so this commit also
removes the abstractmethod annotation and removes the method from
those devices that had an empty initialize body.