Setting: infer readonly attribute and make it always a callable.
The readonly attribute can be True (for when the setting is always
read only), False (for when the setting is always writable), or a
callable (for cases where it is dependent on device state).

This commit changes it so that readonly is always a callable.  The
other two cases can be infered by the absence of a setter in the
setting (which becomes the default) which is the most common case.