Device: ensure it always either enabled or not, never None.
Having a default of None when it's always meant to be True or False is
messed up.

This commit sets the default state of a device enable to False, which
is then changed when _do_enable and _do_disable are called.  If a
device is ever on a weird state, the getter for the state should
check.