device_server: only set the logging level once.
The logging level on the root logger is set once in __console__, then
in serve_devices, then again on DeviceServer.run.  Once is enough.
This should also happen early in the program so that the rest can be
used in other programs without modifying the level that those other
programs may want to use.

This commit removes the setting of logging level for device-server,
leaving it only on __console__.  This is preparation to have a command
line option that sets the logging level (issue #168).