Move Test devices to new microscope.simulators subpackage.
We want the test devices to be used beyond the testsuite and we want
to have more complex test devices like the stage aware camera.

This commit moves the Test* devices, and renames them Simulated*, into
a new microscope.simulators subpackage.  The old ones are kept for
backwards compatibility.  The addition of the enum settings to the
simulated camera, which really is only for testing, is kept on the
testsuite.  Also, the DummySLM and DummyDSP which we don't have a
corresponding part in Microscope yet and are only for Cockpit are kept
in the testsuite.