serve_devices: don't sys.exit when there's no devices to serve.
Don't check for empty list of devices (and exit with sys.exit) in the
middle of the serve_devices.  Instead, let the rest of the code handle
it by simply exiting when there's no more device servers running.  Do
still print a warning though.  We already have a test unit for this
case.