device_server: only inject the "index" kwarg for floating devices.
The "index" kwarg is injected by device_server to be used by SDKs of
floating devices.  It is then taken by the base Device ABC.  This is a
leak of the device_server into the ABCs.  Only floating devices make
use of this "index" (PVCamera, AndorAtmcd, AndorSDK3), so limit this
to FloatingDeviceMixin.

This commit changes so that the index kwarg is not injected unless
device to be served is a FloatingDeviceMixin, adds an initialiser to
FloatingDeviceMixin to take the "index" kwarg, and removes the index
kwarg from the Device ABC.