maint: remove logo from repo and use the one in site on README file.
Ilan really wants an icon to appear when someone checks the source in
github.  The only way to do that is via the README file which is just
a text file.

While nothing on this repo makes use a logo or icon yet (this might
change later if we add it to the microscope-gui program) we added the
svg to the doc directory.  However, if we specify the svg on the
README it doesn't work, it needs to be a bitmap.  We could covert it
but we are already doing it for the site repo (python-microscope.org)
so reference that one instead and remove the source for the image.