device_server.serve_devices: copy devices to prevent change on caller (#211)
This commit has serve_devices make a deepcopy of the devices argument
so that in case we need to make changes to it, we don't modify the
stuff in the caller.  This commit reverts ef280c340260 which was
another fix for this issue.

The way we support floating devices requires that we inject the index
value on the device configuration.  However, we don't want that to
propagate to the caller which may be reusing that configuration (and
it kinda does it when device configuration is not specified see
discussion on #212).