FilterWheel: remove deprecated position setting.
Getting and setting a filterwheel from the device settings has been
deprecated for a while and the related get/set methods have been
around since forever.  Cockpit now also uses the getter and setter
methods which was the only reason this was not removed before.

This commit removes the position setting.