Camera: remove "transform" setting (#260)
We have `Camera.set_transform` and `Camera.get_transform` methods to
control the client transform.  There is also a readout transform (only
used by PVCamera at the moment).  The transform setting exposes the
merge of the those two.  We want settings to be device specific stuff
and not general camera stuff that should be a ABC method.

This commit removes the transform setting from Camera.