Camera._set_readout_transform: do not set client transform again.
After changing the readout or client transform, the final transform
needs to be updated.  That code was only on `set_transform` (which
updates the client transform) so `_set_readout_transform` called
`set_transform` with the current client transform to get the final
transform update.

This commit moves the update of the final transform to a separate
method so that `_set_readout_transform` can just call it.