Remove check for Windows CE.
Python 3.6 removed support for Windows CE (whatever little support
there was).  This means that `os.name` no longer returns the value of
'ce' and to identify Windows is enough to check for the value of 'nt'.
See also https://bugs.python.org/issue27355 for the original CPython
issue about removing this.

Since we are dependent on Python 3.7, support for this was removed in
Python 3.6, and according to the comments on the CPython issue tracker
Windows didn't actually work on any Python 3 version, we can just
remove the checks for Windows CE.