maint: adopt pyproject.toml for build and drop setup.py
We had half of package information in pyproject.toml duplicated from
setup.py.  We were only using pyproject.toml to configure external
tools and left the build to setup.py.  We really needed setup.py
because it handled the call of sphinx-apidoc before sphinx-build and
our manual injection of extra files in source distribution.

This commit replaces our hack to insert files into source
distributions with using `MANIFEST.in` file.  It replaces our call of
sphinx-apidoc with the sphinx apidoc extension (configuration in
doc/conf.py).  So we no longer need a special setup.py and so we
replace it with a more extensive pyproject.toml.

Documentation on how to build documentation and prepare releases is
updated accordingly.