maint: fix optional dependency on PyQt.
In 6e9c67f7 the microscope.gui module changed to be dependent on PyQt.
This new dependency was mentioned on the NEWS file but by mistake was
listed as PySide2 in the actual list of dependencies (setup.py at the
time).  The reason for the mistake is probably that PyQt is a wrapper
around PySide* or PyQt* whichever is installed, and the initial
version of microscope.gui used PySide2 directly.